RENAME: imagedetails_component.dart -> questdetails_component.dart
