+++ lib/pages/questdetails/view/component/questdetails_component.dart
